% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standings.R
\name{calculate}
\alias{calculate}
\title{Title}
\usage{
calculate(
  teams,
  instructors,
  family_reward = 1,
  order_reward = 3,
  class_reward = 5,
  expert_threshold = 4,
  NT_reward = 1,
  VU_reward = 2,
  EN_reward = 3,
  CR_reward = 4,
  RE_reward = 5
)
}
\arguments{
\item{teams}{a list containing a dataframe for each team, see the README document for the format.}

\item{instructors}{a vector of character strings containing the initials of the teachers.}

\item{family_reward}{an integer indicating the number of points rewarded for every new family}

\item{order_reward}{}

\item{class_reward}{}

\item{expert_threshold}{}

\item{NT_reward}{}

\item{VU_reward}{}

\item{EN_reward}{}

\item{CR_reward}{}

\item{RE_reward}{}
}
\value{
a tibble ready to use in plot_standings().
}
\description{
Title
}
