% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standings.R
\name{calculate_standings}
\alias{calculate_standings}
\title{Calculate standings}
\usage{
calculate_standings(
  teams,
  instructors,
  family_reward = 1,
  order_reward = 3,
  class_reward = 5,
  expert_threshold = 4,
  NT_reward = 1,
  VU_reward = 2,
  EN_reward = 3,
  CR_reward = 4,
  RE_reward = 5
)
}
\arguments{
\item{teams}{a list containing a data.frame for each team, see the README document for the format.}

\item{instructors}{a vector of character strings containing the initials of the teachers.}

\item{family_reward}{an integer indicating the number of points rewarded for every new family encountered}

\item{order_reward}{an integer indicating the number of points rewarded for every new order encountered}

\item{class_reward}{an integer indicating the number of points rewarded for every new class encountered}

\item{expert_threshold}{The number of species within a family that has to be encountered for each species to start counting double. set this too a very high number if the feature is not desired.}

\item{NT_reward}{an integer indicating the number of points rewarded when encountering a species listed as Near Threatened on the regional IUCN list.}

\item{VU_reward}{an integer indicating the number of points rewarded when encountering a species listed as Vulnerable on the regional IUCN list.}

\item{EN_reward}{an integer indicating the number of points rewarded when encountering a species listed as Endangered on the regional IUCN list.}

\item{CR_reward}{an integer indicating the number of points rewarded when encountering a species listed as Critically endangered on the regional IUCN list.}

\item{RE_reward}{an integer indicating the number of points rewarded when encountering a species listed as Regionally Extinct on the regional IUCN list.}
}
\value{
a tibble ready to use in plot_standings().
}
\description{
Calculates the standings in the competition.
}
