% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_standings.R
\name{plot_standings}
\alias{plot_standings}
\title{Make a plot of the standings}
\usage{
plot_standings(x)
}
\arguments{
\item{x}{A tibble prepared by calculate_standings()}
}
\value{
a ggplot2 barplot.
}
\description{
Takes a tibble prepared by calculate_standings() and makes an overview plot of the competition standings.
}
